+++ extensions/simple-dictionary/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Simple Dictionary Changelog
 
+## [Set Default Language] - {PR_MERGE_DATE}
+
+- Set a default language in the command `Preferences`.
+
 ## [Initial Version] - 2025-09-16